pyproject.toml
[project]
name = "gitsplain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.80.10",
    "loguru>=0.7.3",
    "openai>=1.0.0",
    "pydantic>=2.12.5",
    "pygithub>=2.5.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
    "ty>=0.0.4",
]
